services/inference_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
boto3==1.35.71
pydantic==2.10.3
python-dotenv==1.0.1
